+++ mcs/class/System.Core/System.Core_test.dll.sources
@@ -61,11 +61,11 @@ System.Linq.Expressions/ExpressionTest_TypeIs.cs
 System.Linq.Expressions/ExpressionTest_UnaryPlus.cs
 System.Linq.Expressions/ExpressionTest_Utils.cs
 System.Threading/ReaderWriterLockSlimTest.cs
-../../corlib/Test/System.Security.Cryptography/HashAlgorithmTest.cs
-../../corlib/Test/System.Security.Cryptography/SHA1Test.cs
-../../corlib/Test/System.Security.Cryptography/SHA256Test.cs
-../../corlib/Test/System.Security.Cryptography/SHA384Test.cs
-../../corlib/Test/System.Security.Cryptography/SHA512Test.cs
+../../corlib/Test/System.Security.Cryptography/HashAlgorithmTestBase.cs
+../../corlib/Test/System.Security.Cryptography/SHA1TestBase.cs
+../../corlib/Test/System.Security.Cryptography/SHA256TestBase.cs
+../../corlib/Test/System.Security.Cryptography/SHA384TestBase.cs
+../../corlib/Test/System.Security.Cryptography/SHA512TestBase.cs
 System.Security.Cryptography/AesCryptoServiceProviderTest.cs
 System.Security.Cryptography/AesManagedTest.cs
 System.Security.Cryptography/CngAlgorithmTest.cs
