+++ mcs/class/System.Core/Test/System.Security.Cryptography/SHA1CngTest.cs
@@ -25,7 +25,7 @@ namespace MonoTests.System.Security.Cryptography {
 	// same results (hence should run a common set of unit tests).
 
 	[TestFixture]
-	public class SHA1CngTest : SHA1Test {
+	public class SHA1CngTest : SHA1TestBase {
 
 		[SetUp]
 		public override void SetUp ()
