+++ mcs/class/System.Core/Test/System.Security.Cryptography/SHA512CngTest.cs
@@ -25,7 +25,7 @@ namespace MonoTests.System.Security.Cryptography {
 	// same results (hence should run a common set of unit tests).
 
 	[TestFixture]
-	public class SHA512CngTest : SHA512Test {
+	public class SHA512CngTest : SHA512TestBase {
 
 		[SetUp]
 		public override void SetUp ()
