+++ mcs/class/corlib/Test/System.Security.Cryptography/SHA256ManagedTest.cs
@@ -23,7 +23,7 @@ namespace MonoTests.System.Security.Cryptography {
 // same results (hence should run a common set of unit tests).
 
 [TestFixture]
-public class SHA256ManagedTest : SHA256Test {
+public class SHA256ManagedTest : SHA256TestBase {
 
 	[SetUp]
 	public override void SetUp ()
