+++ mcs/class/corlib/Test/System.Security.Cryptography/SHA384ManagedTest.cs
@@ -23,7 +23,7 @@ namespace MonoTests.System.Security.Cryptography {
 // same results (hence should run a common set of unit tests).
 
 [TestFixture]
-public class SHA384ManagedTest : SHA384Test {
+public class SHA384ManagedTest : SHA384TestBase {
 
 	[SetUp]
 	public override void SetUp ()
