+++ mcs/class/corlib/Test/System.Security.Cryptography/SHA512ManagedTest.cs
@@ -23,7 +23,7 @@ namespace MonoTests.System.Security.Cryptography {
 // same results (hence should run a common set of unit tests).
 
 [TestFixture]
-public class SHA512ManagedTest : SHA512Test {
+public class SHA512ManagedTest : SHA512TestBase {
 
 	[SetUp]
 	public override void SetUp ()
