+++ mcs/class/corlib/corlib_test.dll.sources
@@ -261,7 +261,8 @@ System.Security.Cryptography/DSASignatureDeformatterTest.cs
 System.Security.Cryptography/DSASignatureFormatterTest.cs
 System.Security.Cryptography/DSATest.cs
 System.Security.Cryptography/FromBase64Transform.cs
-System.Security.Cryptography/HashAlgorithmTest.cs
+System.Security.Cryptography/HashAlgorithmTestBase.cs
+System.Security.Cryptography/HashAlgorithmTestImpl.cs
 System.Security.Cryptography/HMACMD5Test.cs
 System.Security.Cryptography/HMACRIPEMD160Test.cs
 System.Security.Cryptography/HMACSHA1Test.cs
@@ -294,13 +295,17 @@ System.Security.Cryptography/RSAPKCS1SignatureDeformatterTest.cs
 System.Security.Cryptography/RSAPKCS1SignatureFormatterTest.cs
 System.Security.Cryptography/RSATest.cs
 System.Security.Cryptography/SHA1CryptoServiceProviderTest.cs
-System.Security.Cryptography/SHA1Test.cs
+System.Security.Cryptography/SHA1TestBase.cs
+System.Security.Cryptography/SHA1TestImpl.cs
 System.Security.Cryptography/SHA256ManagedTest.cs
-System.Security.Cryptography/SHA256Test.cs
+System.Security.Cryptography/SHA256TestBase.cs
+System.Security.Cryptography/SHA256TestImpl.cs
 System.Security.Cryptography/SHA384ManagedTest.cs
-System.Security.Cryptography/SHA384Test.cs
+System.Security.Cryptography/SHA384TestBase.cs
+System.Security.Cryptography/SHA384TestImpl.cs
 System.Security.Cryptography/SHA512ManagedTest.cs
-System.Security.Cryptography/SHA512Test.cs
+System.Security.Cryptography/SHA512TestBase.cs
+System.Security.Cryptography/SHA512TestImpl.cs
 System.Security.Cryptography/SignatureDescriptionTest.cs
 System.Security.Cryptography/SymmetricAlgorithmTest.cs
 System.Security.Cryptography/SymmetricAlgorithm2Test.cs
